requirements.txt.txt
aiogram==3.18.0
python-dotenv==1.0.0